requirements.txt
fastapi
pandas
numpy
scikit-learn
joblib
uvicorn[standard]
pydantic

# Exporter for Google Cloud
opentelemetry-exporter-gcp-trace
